requirements.txt
pynetgear-enhanced>=0.1.1
pyYAML>=5.1
requests>=2.21.0
